patina::utils
Enum Error

pub enum Error {
    FileRead(PathBuf, Error),
    TomlParse(Error),
    RenderTemplate(RenderError),
}

An enum representing all possible errors that can occur in this crate

Variants

FileRead(PathBuf, Error)

An error that occurs when a file cannot be read

TomlParse(Error)

An error that occurs when parsing Toml data

RenderTemplate(RenderError)

An error that occurs when rendering a handlebars template

Implementations

impl Error

pub fn is_file_read(&self) -> bool

Returns true if this is a Error::FileRead, otherwise false

pub fn as_file_read_mut(&mut self) -> Option<(&mut PathBuf, &mut Error)>

Optionally returns mutable references to the inner fields if this is a Error::FileRead, otherwise None

pub fn as_file_read(&self) -> Option<(&PathBuf, &Error)>

Optionally returns references to the inner fields if this is a Error::FileRead, otherwise None

pub fn into_file_read(self) -> Result<(PathBuf, Error), Self>

Returns the inner fields if this is a Error::FileRead, otherwise returns back the enum in the Err case of the result

pub fn is_toml_parse(&self) -> bool

Returns true if this is a Error::TomlParse, otherwise false

pub fn as_toml_parse_mut(&mut self) -> Option<&mut Error>

Optionally returns mutable references to the inner fields if this is a Error::TomlParse, otherwise None

pub fn as_toml_parse(&self) -> Option<&Error>

Optionally returns references to the inner fields if this is a Error::TomlParse, otherwise None

pub fn into_toml_parse(self) -> Result<Error, Self>

Returns the inner fields if this is a Error::TomlParse, otherwise returns back the enum in the Err case of the result

pub fn is_render_template(&self) -> bool

Returns true if this is a Error::RenderTemplate, otherwise false

pub fn as_render_template_mut(&mut self) -> Option<&mut RenderError>

Optionally returns mutable references to the inner fields if this is a Error::RenderTemplate, otherwise None

pub fn as_render_template(&self) -> Option<&RenderError>

Optionally returns references to the inner fields if this is a Error::RenderTemplate, otherwise None

pub fn into_render_template(self) -> Result<RenderError, Self>

Returns the inner fields if this is a Error::RenderTemplate, otherwise returns back the enum in the Err case of the result